BotTesting/requirements.txt
appium-python-client
selenium